it works but not pretty
