+++ client/src/Components/Auth/signin.js
@@ -1,12 +1,10 @@
 import React, { Component } from 'react'
-import './signin.css'
-import Modal from 'react-bootstrap/Modal'
-import Button from 'react-bootstrap/Button'
-
-class SignIn extends Component {
+class SignUp extends Component {
     state = {
         email: '',
         password: '',
+        firstName: '',
+        lastName: '',
     }
     handleChange = e => {
         this.setState({
@@ -17,25 +15,52 @@ class SignIn extends Component {
         e.preventDefault()
         console.log(this.state)
     }
-
     render() {
         return (
-            <Modal.Dialog>
-                <Modal.Header closeButton>
-                    <Modal.Title>Modal title</Modal.Title>
-                </Modal.Header>
-
-                <Modal.Body>
-                    <p>Modal body text goes here.</p>
-                </Modal.Body>
-
-                <Modal.Footer>
-                    <Button variant="secondary">Close</Button>
-                    <Button variant="primary">Save changes</Button>
-                </Modal.Footer>
-            </Modal.Dialog>
+            <div className="container">
+                <form className="white" onSubmit={this.handleSubmit}>
+                    <h5 className="grey-text text-darken-3">Sign Up</h5>
+                    <div className="input-field">
+                        <label htmlFor="email">Email</label>
+                        <input
+                            type="email"
+                            id="email"
+                            onChange={this.handleChange}
+                        />
+                    </div>
+                    <div className="input-field">
+                        <label htmlFor="password">Password</label>
+                        <input
+                            type="password"
+                            id="password"
+                            onChange={this.handleChange}
+                        />
+                    </div>
+                    <div className="input-field">
+                        <label htmlFor="firstName">First Name</label>
+                        <input
+                            type="text"
+                            id="firstName"
+                            onChange={this.handleChange}
+                        />
+                    </div>
+                    <div className="input-field">
+                        <label htmlFor="lastName">Last Name</label>
+                        <input
+                            type="text"
+                            id="lastName"
+                            onChange={this.handleChange}
+                        />
+                    </div>
+                    <div className="input-field">
+                        <button className="btn pink lighten-1 z-depth-0">
+                            Sign Up
+                        </button>
+                    </div>
+                </form>
+            </div>
         )
     }
 }
 
-export default SignIn
+export default SignUp
