+++ client/src/Constants/Firebase/context.js
@@ -2,9 +2,9 @@ import React from 'react'
 const FirebaseContext = React.createContext(null)
 
 export const withFirebase = Component => props => (
-    <FirebaseContext.Consumer>
-        {firebase => <Component {...props} firebase={firebase} />}
-    </FirebaseContext.Consumer>
-)
+  <FirebaseContext.Consumer>
+    {firebase => <Component {...props} firebase={firebase} />}
+  </FirebaseContext.Consumer>
+);
 
 export default FirebaseContext
