+++ package.json
@@ -59,7 +59,8 @@
     "prettier": "^1.18.2",
     "pretty-quick": "^1.11.1",
     "react-datepicker": "^2.8.0",
-    "react-number-format": "^4.2.0"
+    "react-number-format": "^4.2.0",
+    "recompose": "^0.30.0"
   },
   "repository": {
     "type": "git",
